pyproject.toml
[project]
name = "ontogent"
version = "0.1.0"
description = "AI agent for finding suitable UBERON ontology terms"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "anthropic>=0.6.0",
    "langchain>=0.0.267",
    "pydantic>=2.0.0",
    "pytest>=7.3.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.0.278",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.3.1",
    "black>=23.7.0",
    "ruff>=0.0.278",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
ignore = []
